feat(add):
Improves the application's interface and usability

Updates `App.xaml` to include resource dictionaries.

Modifies `App.xaml.cs` to set the `MainPage` to
`ContractedEvent` and adds the `CreateWindow` method to
configure the window.

Adds "Roboto-Regular" and "Roboto-Bold" fonts to
`MauiProgram.cs`. Updates colors in `Colors.xaml`,
including new values ​​for `Primary`, `PrimaryDark`
and a new `background` background color.

Reworks `ContractedEvent.xaml` with an intuitive
layout and input fields for event information, as well as
a button for contracting. Updates `ContractedEvent.xaml`
with the `Button_Clicked` method for navigation and
exception handling.

Adds `bg.svg` with a new SVG background that improves the
aesthetics of the interface.